requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.9.0
click==8.1.8
dnspython==2.7.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
pytz==2022.7.1
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
